feat(cli): Submit release listens using an end time parameter
Start and end time options are mutually exclusive.
The new default behavior is to use the current time as end time, which
makes more sense than using it as the start time (as before).
With this new behavior, the start time becomes mandatory for the
submission of a single track where only title and artist are known.